requirements.txt
Django==4.2
djangorestframework==3.15.1
openpyxl==3.0.7
PyPDF2==3.0.0
django-redis==5.3.0
pandas == 2.2.1
psycopg2-binary == 2.9.8
mysqlclient == 2.2.3
cachetools == 5.3.2